Revert "Salesforce attachment changes" (#1782)
* Revert "Respond 503 in correspondence with EVSS GIBS scheduled uptime (#1769)"

This reverts commit 16b469a9625a8b8a62cc9ecffc675cca2f22fc43.

* Revert "Pension burial new server (#1779)"

This reverts commit bd36471fdaf8d7fdbc1bd6d70a3b899d3dc0e172.

* Revert "vets.gov -> Vets.gov (#1778)"

This reverts commit 8b292f20e39552452eb0f3ebea0d9140a305960c.

* Revert "Change timezone to Central (#1776)"

This reverts commit 50a5fc77ea196cbd8d5f09d7766252279fc14593.

* Revert "Salesforce attachment changes (#1773)"

This reverts commit db1f63739d1055ad7d87cf01e4f2cce06b0a5c0a.

+++ lib/vic/service.rb
@@ -28,6 +28,9 @@ module VIC
       phone
     ].freeze
 
+    class AttachmentUploadFailed < StandardError
+    end
+
     def oauth_params
       {
         grant_type: 'urn:ietf:params:oauth:grant-type:jwt-bearer',
@@ -90,23 +93,22 @@ module VIC
       file_name = "#{description}.#{mime_type.split('/')[1]}"
       file_path = Common::FileHelpers.generate_temp_file(file_body)
 
-      content_version_id = client.create!(
-        'ContentVersion',
-        Title: description, PathOnClient: file_name,
-        VersionData: Restforce::UploadIO.new(
+      success = client.create(
+        'Attachment',
+        ParentId: case_id,
+        Name: file_name,
+        Body: Restforce::UploadIO.new(
           file_path,
           mime_type
         )
       )
-
-      client.create!(
-        'ContentDocumentLink',
-        ContentDocumentId: client.find('ContentVersion', content_version_id)['ContentDocumentId'],
-        ShareType: 'V', LinkedEntityId: case_id
-      )
-
       File.delete(file_path)
-      form_attachment.destroy
+
+      if success
+        form_attachment.destroy
+      else
+        raise AttachmentUploadFailed
+      end
     end
 
     def get_attachment_records(form)
